+++ messages/ru/app.php
@@ -412,4 +412,15 @@ return
     'Incoming goods'                                            => 'Поступление товара',
     "We've got the goods interesting you"                       => 'У нас появился интересующий Вас товар',
     "Product name"                                              => 'Название товара',
+    "Discount coupons"                                          => 'Скидочные купоны',
+    "Coupon"                                                    => 'Купон',
+    "Active from"                                               => 'Активность от',
+    "Active to"                                                 => 'Активность до',
+    "Maximum number of performances"                            => 'Максимальное число использований',
+
+
+
+    "Not set coupon code"                                       => 'Не указан код купона',
+    "Coupon does not exist or is not active"                    => 'Купон не существует или не активен',
+    "Coupon successfully installed"                             => 'Купон успешно установлен',
 ];