+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Changelog
 
 1.0.0-alpha27.pre
 -----------------
+ * Implementation of the coupon codes
+ * Update js api
+ * Recalculate() cart
+ * Removed \skeeks\cms\widgets\rbac\PermissionForRoles
+ * Fixed bugs
  * Search and filtering the profiles of buyers
  * ShopNoAuthCheckoutWidget removed -> \skeeks\cms\shopCheckoutSimple\ShopCheckoutSimpleWidget
  * ShopCartStepsWidget removed -> \skeeks\cms\shopCartStepsWidget\ShopCartStepsWidget